I know how to mod the map
It's all about save format

province.bmp is the easiest one you just have to confirm you're no using a duplicate color
terrain in .bmp file and 8 bits. only use the colors in the file.
topology.bmp has been the hardest to modify. the trick is to save in true grayscale. you can do that on GIMP (as in, edit the file in your favorite program, then use gimp to give the proper file format

I still don't know if I'm modifying the colormaps.dds correctly