Province bug fixed
The way sea zones are read in default.map file considers that EVERY province ID between the two listed ID in a specific sea zone are either adjacent or, if not, are considered a lake. Since I wanted lake nichupte to be a lake, I thought it was easy to put into the lakes sea zones. that's what created or issue.

Fortunately it's fixed now.